Make  getBacklightProtocolVersion compatible with all versions
